removed device manager
The Device Manager. This is a module that handles active and bonded
connections. Since this tutorial is about advertising and not
connections it is not really relevant. Still I chose to leave it in the
code as it is extensively used in more advanced examples, so I figured
you might just as well get used to it being there and it doesn't do any
harm.